feat: expose alternate-file parsing via `alternate::parse()`
This is mostly for completeness and more low-level access.
